First working version of the low-memory algorithm

+++ benchmarks/zipfian.h
@@ -27,7 +27,7 @@ public:
 
   uint64_t operator () (size_t i) {
     uint64_t r = parlay::hash64(i);
-    double u = ((double) r)/std::numeric_limits<uint64_t>::max(); // uniform between 0.0 and 1.0
+    double u = static_cast<double>(r) / static_cast<double>(std::numeric_limits<uint64_t>::max()); // uniform between 0.0 and 1.0
     double uz = u * zeta_n_;
     if (uz < 1.0) return 0;
     if (uz < 1.0 + std::pow(0.5, theta_)) return 1;
